+++ core/src/xmake/engine.c
@@ -182,8 +182,9 @@ tb_int_t xm_path_is_absolute(lua_State* lua);
 
 // the hash functions
 tb_int_t xm_hash_uuid4(lua_State* lua);
-tb_int_t xm_hash_sha256(lua_State* lua);
+tb_int_t xm_hash_sha(lua_State* lua);
 tb_int_t xm_hash_md5(lua_State* lua);
+tb_int_t xm_hash_xxhash(lua_State* lua);
 
 // the base64 functions
 tb_int_t xm_base64_encode(lua_State* lua);
@@ -406,8 +407,9 @@ static luaL_Reg const g_path_functions[] =
 static luaL_Reg const g_hash_functions[] =
 {
     { "uuid4",          xm_hash_uuid4  }
-,   { "sha256",         xm_hash_sha256 }
+,   { "sha",            xm_hash_sha    }
 ,   { "md5",            xm_hash_md5    }
+,   { "xxhash",         xm_hash_xxhash }
 ,   { tb_null,          tb_null        }
 };